fix and form (converter): moving some functions in parser.py and other in checker.py. Removing the function from __main__ and defining main() function calling parser and checker
